requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
httpx==0.28.1
gspread==6.2.0
google-auth==2.38.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.8.1
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
